RaspberryPi HAT for controlling 2 DC motors and 4 servos. There is also 8 general outputs and 8 general inputs. Not yet complete.
This HAT is intended for a Lego robot project. For DC motors, bot Lego motors and other motors can be used, and for servos, ordinary RC servos using PWM control signals can be used. The servos must then be adapted to fit Lego. The HAT is provied with 12V supply voltage from a battery, and power the Raspberry PI, as well as the servos using a 5V DC-DC converter.